Delete f
The following code is a password generator with four options:
    alphabet
    number+alphabet+symbols
    number+alphabet+symbols+foreign characters
    all valid bits
Provided is a file called output.txt, which should also be downloaded to the same directory as the file as it allows the output to be stored.
Provided also is launch.json, which is a set of launch instructions for Microsoft visual studio. it contains the correct character encoding and command line arguments to see utf-8 characters. Without this, settings 3 and 4 wont work and will only show '?' in the command prompt. Other IDE's might be compatible though.