(Kevin) Constrained UI, Volley, Error Catching
The Ui now doesn't blast off the page every time you get a response. Also, it catches a few errors instead of showing no response. Changed the API key to the Dogecoin testnet for demo purposes (BTC Testnet addresses are full). HTTP requests are done with Volley instead of AsyncTask which is... better?